Fixed Sanselan bug when the IPTC segment contains invalid tags
Shade the IPTCParser Sanselan class.

Indeed, this parser returns the IPTC tags as soon as an invalid record is found. Unfortunately, if the IPTC metadata are edited using Preview of Picasa, some weird records are added to the segments, and so keywords are not found.

The fix is really simple, I just skip the invalid records.

It seems that XMP metadata re more reliable, but require more work.